finished small poc function for showing yubikeyctr
This small POC shows the Yubikey signature counter via barista.
It works, but I will not use it. The pressure on the CPU is too high,
even when I call the `gpg --card-status` command only every 5 seconds.
Also it will trigger the yubikey every 5 seconds and trigger the LED in
it. Not sure if it so good, to constantly trigger it.

Signed-off-by: Christian Rebischke <chris@nullday.de>